Clean up the message and use the new guard for detecting parsing tests.

+++ Bluepill-cli/Bluepill-cli/Simulator/SimulatorMonitor.m
@@ -219,15 +219,12 @@ typedef NS_ENUM(NSInteger, SimulatorState) {
     // Timeout or crash on a test means we should skip it when we rerun the tests
     [self updateExecutedTestCaseList:testName inClass:testClass];
 
-    if (![[self.device stateString] isEqualToString:@"Shutdown"]) {
-        // self.appPID can be zero when running the parsing tests
-        // since we're not actually creating a simulator and running an app.
+    if (![[self.device stateString] isEqualToString:@"Shutdown"] && !self.config.testing_NoAppWillRun) {
         [BPUtils printInfo:ERROR withString:@"Will kill the process with appPID: %d", self.appPID];
-        if (self.appPID && (kill(self.appPID, 0) == 0) && (kill(self.appPID, SIGTERM) < 0)) {
-            [BPUtils printInfo:ERROR withString:@"Failed to kill the process with appPID: %d", self.appPID];
-            perror("kill");
-        } else {
-            [BPUtils printInfo:ERROR withString:@"Success killing the process with appPID: %d", self.appPID];
+        NSAssert(self.appPID > 0, @"Failed to find a valid PID");
+        if ((kill(self.appPID, 0) == 0) && (kill(self.appPID, SIGTERM) < 0)) {
+            [BPUtils printInfo:ERROR withString:@"Failed to kill the process with appPID: %d: %s",
+                self.appPID, strerror(errno)];
         }
     }
